Move `PartitionId` to `Vocabulary/` and use it in bootloader

+++ Sts1CobcSw/Firmware/RfCommunicationThread.cpp
@@ -681,7 +681,8 @@ auto Handle(UpdateEduQueueFunction const & function, RequestId const & requestId
     {
         edu::programQueue.PushBack(entry);
     }
-    DEBUG_PRINT("Updated EDU queue with %u entries\n", function.queueEntries.size());
+    DEBUG_PRINT("Updated EDU queue with %d entries\n",
+                static_cast<int>(function.queueEntries.size()));
     SendAndWait(SuccessfulCompletionOfExecutionVerificationReport(requestId));
     ResumeEduProgramQueueThread();
 }
